==> 02_programs/01a_download_crimedata.Rmd <==
---
title: "Downloads crime data from NYC Open Data"
output: html_document
author: "Adelaide Currin, Haowei Wang, Jennah Gosciak, Yiping Zuo"
date: "2022-11-07"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(sf)
```

# Load arrests data

```{r}
# load arrest data
# note: this may take some time!
ar <- read_csv("https://data.cityofnewyork.us/resource/8h9b-rp9u.csv?$where=arrest_date>='2017-01-01'+AND+arrest_date<'2022-01-01'&$limit=10000000")
ar

ar %>%
  summarize(across(arrest_date, .fns = lst(min, max)))

ar %>%
  write_csv("../00_data/arrests_2017_2020.csv", na = "")
```

```{r}
# Breaking arrests into categories (felonies and misdemeanor). Robbery, assault, burglary, weapons, misdemeanor offenses (e.g., criminal mischief, fraud, gambling, loitering, petty theft, and larceny), other felonies (e.g., escape 3 and forgery), drugs (e.g., dangerous drugs), property (e.g., grand larceny, burglary, and burglary tools), and violent felonies (e.g., homicide, rape, robbery, arson, felony assault, and kidnapping). Then also https://www.criminaljustice.ny.gov/crimnet/ojsa/stats.htm

ar <- ar %>%
  filter(law_cat_cd == "M" | law_cat_cd == "F") %>%
  mutate(violent_ellen = ifelse(ofns_desc %in%
    c("RAPE", "ROBBERY", "FELONY ASSAULT", "MURDER & NON-NEGL. MANSLAUGHTER", "MURDER & NON-NEGL. MANSLAUGHTE", "HOMICIDE-NEGLIGENT-VEHICLE", "HOMICIDE-NEGLIGENT,UNCLASSIFIED"), 1, 0)) %>%
  mutate(property_ellen = ifelse(ofns_desc %in%
    c("ARSON", "BURGLARY", "GRAND LARCENY", "GRAND LARCENY OF MOTOR VEHICLE", "PETIT LARCENY", "THEFT-FRAUD"), 1, 0))

ar %>%
  write_csv("../00_data/arrests_2017_2020.csv", na = "")
```


```{r}
# average arrests per day
ar %>%
  group_by(arrest_date) %>%
  summarize(n = n()) %>%
  summarize(avg_pday = mean(n))

ar %>%
  mutate(arrest_month_yr = as.Date(str_c(month(arrest_date), 1, year(arrest_date), sep = "-"), "%m-%d-%Y")) %>%
  group_by(arrest_month_yr, law_cat_cd) %>%
  summarize(n = n()) %>%
  ggplot() +
  geom_line(aes(x = arrest_month_yr, y = n, color = law_cat_cd)) +
  labs(x = "Month/year", y = "Number of arrests per month", color = "Arrest type") +
  theme_classic()
```

# Load complaints data

```{r}
cp <- read_csv("https://data.cityofnewyork.us/resource/qgea-i56i.csv?$where=cmplnt_fr_dt>='2017-01-01'+AND+cmplnt_fr_dt<'2022-01-01'&$limit=10000000")
cp

cp %>%
  summarize(across(cmplnt_fr_dt, .fns = lst(min, max)))

unique(cp$ofns_desc)

cp <- cp %>%
  filter(law_cat_cd == "FELONY" | law_cat_cd == "MISDEMEANOR") %>%
  mutate(violent_ellen = ifelse(ofns_desc %in%
    c("RAPE", "ROBBERY", "FELONY ASSAULT", "MURDER & NON-NEGL. MANSLAUGHTER", "MURDER & NON-NEGL. MANSLAUGHTE", "HOMICIDE-NEGLIGENT-VEHICLE", "HOMICIDE-NEGLIGENT,UNCLASSIFIED"), 1, 0)) %>%
  mutate(property_ellen = ifelse(ofns_desc %in%
    c("ARSON", "BURGLARY", "GRAND LARCENY", "GRAND LARCENY OF MOTOR VEHICLE", "PETIT LARCENY", "THEFT-FRAUD"), 1, 0))


cp %>%
  write_csv("../00_data/complaints_2017_2020.csv", na = "")
```

# Load OZ data
```{r}
# opportunity zones
oz <- st_read("../00_data/Opportunity_Zone_Eligible_Census_Tracts.geojson")
oz %>%
  head()
```

```{r}
# filter by state
ny_oz <- oz %>%
  filter(!is.na(DESIGNATED)) %>%
  filter(STATE == 36)

nyc_oz <- ny_oz %>%
  filter(COUNTY %in% c("005", "061", "085", "081", "047"))

nyc_oz %>%
  st_write("../00_data/nyc_oz.geojson",
    delete_dsn = TRUE
  )

nyc_oz %>%
  as.data.frame() %>%
  group_by(DESIGNATED) %>%
  summarize(n = n())
```
