==> 02_programs/01c_oz_boundary.Rmd <==
---
title: "Descriptive Analysis of OZs using ACS Census Data"
author: "Adelaide Currin, Haowei Wang, Jennah Gosciak, Yiping Zuo"
date: "11-18-2022"
output:
  html_document:
    df_print: paged
---

Creates a buffer based on the average square footage of a NYC block.

```{r, include=F}
library(tidyverse)
library(tidycensus)
library(magrittr)
library(sf)
library(wesanderson) # for colors :)
library("lmtest")
library("sandwich")
library(readxl)
```
```{r}
# load ozs for nyc
oz <- st_read("../00_data/nyc_oz.geojson")

boros <- st_read("https://data.cityofnewyork.us/resource/7t3b-ywvw.geojson")

oz %>%
  filter(is.na(DESIGNATED)) %>%
  nrow() %>%
  equals(0) %>%
  stopifnot()

oz_proj <- oz %>%
  # project to nys plane in feet
  st_transform(2263)
oz_proj

oz_bound <- oz_proj %>%
  # we only want boundaries between selected and eligible tracts
  # need to filter that
  filter(DESIGNATED == "True") %>%
  st_union() %>%
  st_boundary()

oz_bound %>%
  st_union() %>%
  st_write("../00_data/oz_bound.geojson", delete_dsn = TRUE)
```
