==> 02_programs/02a_spatial_join.Rmd <==
---
title: "Spatial join between NYC OZ Boundary and crime incidents"
output: html_document
author: "Adelaide Currin, Haowei Wang, Jennah Gosciak, Yiping Zuo"
date: "2022-11-07"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lubridate)
library(sf)
library(magrittr)
library(stats)
```

## Load data

```{r}
# load oz data
nyc_oz <- st_read("../00_data/nyc_oz.geojson") %>%
  # convert to boolean
  mutate(DESIGNATED = if_else(DESIGNATED == "True", TRUE, FALSE)) %>%
  # convert to state plan coordinate system, feet
  st_transform(2263)
```

```{r, message = F}
# load census tracts and census blocks
cb <- st_read("https://data.cityofnewyork.us/resource/twhy-dzjp.geojson?$limit=1000000") %>%
  st_transform(2263) %>%
  mutate(
    county_code = case_when(
      boro_code == 1 ~ "36061",
      boro_code == 2 ~ "36005",
      boro_code == 3 ~ "36047",
      boro_code == 4 ~ "36081",
      boro_code == 5 ~ "36085"
    ),
    GEOID = str_c(county_code, ct2010),
    GEOID_b = str_c(county_code, ct2010, cb2010)
  ) %>%
  mutate(geometry = st_make_valid(geometry))

table(st_is_valid(cb))

cb <- cb %>%
  left_join(nyc_oz %>%
    as.data.frame() %>%
    mutate(GEOID = str_trim(GEOID)) %>%
    select(-geometry), on = "GEOID")

cb %>%
  as.data.frame() %>%
  distinct(DESIGNATED)

cb %>%
  mutate(DESIGNATED_lab = case_when(
    DESIGNATED == T ~ "Designated",
    DESIGNATED == F ~ "Eligible, not designated",
    TRUE ~ "Not eligible"
  )) %>%
  st_write("../00_data/census_blocks_labeled.geojson", delete_dsn = T)
```

```{r}
# load census tracts and census blocks
ct <- st_read("https://data.cityofnewyork.us/resource/i69b-3rdj.geojson?$limit=1000000") %>%
  st_transform(2263) %>%
  mutate(
    county_code = case_when(
      boro_code == 1 ~ "36061",
      boro_code == 2 ~ "36005",
      boro_code == 3 ~ "36047",
      boro_code == 4 ~ "36081",
      boro_code == 5 ~ "36085"
    ),
    GEOID = str_c(county_code, ct2010),
  ) %>%
  mutate(geometry = st_make_valid(geometry))
ct

table(st_is_valid(ct))

ct <- ct %>%
  left_join(nyc_oz %>%
    as.data.frame() %>%
    mutate(GEOID = str_trim(GEOID)) %>%
    select(-geometry), on = "GEOID")

ct %>%
  as.data.frame() %>%
  distinct(DESIGNATED)

ct %>%
  mutate(DESIGNATED_lab = case_when(
    DESIGNATED == T ~ "Designated",
    DESIGNATED == F ~ "Eligible, not designated",
    TRUE ~ "Not eligible"
  )) %>%
  st_write("../00_data/census_tracts_labeled.geojson", delete_dsn = T)

eeptools::isid(as.data.frame(ct), "GEOID")
```

```{r}
# load acs data
acs_oz <- read_csv("../00_data/acs_oz.csv")
acs_bg <- read_csv("../00_data/acs_bg.csv")
```

```{r}
# load the arrests
ar <- read_csv("../00_data/arrests_2017_2020.csv")
ar %>%
  head()

ar %>%
  filter(arrest_date < mdy("06-14-2018") & arrest_date >= mdy("06-14-2017")) %>%
  nrow()

ar_sf <- ar %>%
  st_as_sf(coords = c("longitude", "latitude"), crs = 4326) %>%
  # convert to projected coordinate system, state plan ny, feet
  st_transform(2263)

ar_sf
```

```{r}
cp <- read_csv("../00_data/complaints_2017_2020.csv")
cp %>%
  head()

cp %>%
  filter(cmplnt_fr_dt < mdy("06-14-2018") & cmplnt_fr_dt >= mdy("06-14-2017")) %>%
  nrow()

cp_sf <- cp %>%
  filter(!is.na(longitude) & !is.na(latitude)) %>%
  st_as_sf(coords = c("longitude", "latitude"), crs = 4326) %>%
  # convert to projected coordinate system, state plan ny, feet
  st_transform(2263)
```

```{r}
nyc_oz %>%
  ggplot() +
  geom_sf(aes(fill = DESIGNATED)) +
  theme_void()
```

```{r}
oz_bound <- st_read("../00_data/oz_bound.geojson")
oz_bound %>%
  head()

oz_bound %>%
  ggplot() +
  geom_sf() +
  theme_void()
```

## Performing spatial  joins

```{r}
# spatial joins
ar_oz <- ar_sf %>%
  st_join(cb, left = TRUE) %>%
  mutate(distance_boundary = st_distance(ar_sf, oz_bound))

ar_oz %>%
  as.data.frame() %>%
  group_by(DESIGNATED) %>%
  summarize(n = n())

ar_oz %>%
  head()

stopifnot(nrow(as.data.frame(ar_sf)) == nrow(as.data.frame(ar_oz)))
```

```{r}
ar_oz %>%
  mutate(distance_boundary = as.numeric(distance_boundary)) %>%
  filter(!is.na(distance_boundary)) %>%
  filter(distance_boundary <= 1) %>%
  ggplot() +
  geom_sf(color = "blue", alpha = 0.01) +
  geom_sf(data = oz_bound, color = "red") +
  theme_void()

ar_oz %>%
  mutate(
    distance_boundary = as.numeric(distance_boundary),
    distance_boundary = round(distance_boundary, digits = -1)
  ) %>%
  as.data.frame() %>%
  filter(DESIGNATED %in% c(TRUE, FALSE), distance_boundary < 1000) %>%
  filter(DESIGNATED == F) %>%
  group_by(distance_boundary, DESIGNATED) %>%
  summarize(n = n())
```

```{r}
# calculate centroid of census blocks
cb_centroid <- cb %>%
  st_centroid()

ggplot() +
  geom_sf(data = cb_centroid, color = "gray") +
  geom_sf(data = oz_bound, color = "blue") +
  theme_void()
```

```{r}
# compute distances
centroid_distances <- cb_centroid %>%
  mutate(centroid_distance_boundary = st_distance(cb_centroid, oz_bound))

centroid_distances %>%
  st_write("../03_output/centroid_distances.geojson", delete_dsn = T)
```

```{r}
# compute pairs
# code is inspired by work from Chen, Glaeser, and Wessel
x <- centroid_distances %>%
  filter(DESIGNATED == T)
y <- centroid_distances %>%
  filter(DESIGNATED == F | is.na(DESIGNATED))

dist_mat <- st_distance(x, y)
```
```{r}
min_dist_index <- y[apply(as.matrix(dist_mat), 1, which.min), ] %>%
  as.data.frame() %>%
  magrittr::extract2("GEOID_b")
min_dist_index

min_dist <- apply(as.matrix(dist_mat), 1, min)

centroid_distance_pairs <- x %>%
  as.data.frame() %>%
  select(GEOID_b) %>%
  mutate(
    dist_pair = min_dist_index,
    dist = min_dist
  )

centroid_distance_pairs %>%
  as.data.frame() %>%
  head()

centroid_distance_pairs %>%
  write_csv("../03_output/centroid_distance_pairs.csv")
```

```{r}
cp_oz <- cp_sf %>%
  st_join(cb, left = TRUE) %>%
  mutate(distance_boundary = st_distance(cp_sf, oz_bound))

cp_oz %>%
  as.data.frame() %>%
  group_by(DESIGNATED) %>%
  summarize(n = n())

cp_oz %>%
  head()
```

```{r}
acs_oz_vars <- acs_oz %>%
  select(GEOID, population:pctover64) %>%
  mutate(GEOID = as.character(GEOID))

cp_oz_acs <- cp_oz %>%
  left_join(acs_oz_vars, by = "GEOID")

ar_oz_acs <- ar_oz %>%
  left_join(acs_oz_vars, by = "GEOID")

stopifnot(nrow(as.data.frame(ar_oz_acs)) == nrow(as.data.frame(ar_oz)))
```

```{r}
# making sure census blocks and OZ tracts line up
checking_rows <- ar_oz_acs %>%
  as.data.frame() %>%
  # note: calculating this at the Census block level
  group_by(
    GEOID_b, GEOID, DESIGNATED, pct_white, pct_higher_ed, pct_rent, pct_native_hc_covered,
    pct_poverty, minutes_commute_avg, dec_score, SE_Flag, vacancyrate, pctunder18, pctover64
  ) %>%
  summarize(across(c(violent_ellen, property_ellen), ~ sum(., na.rm = T))) %>%
  filter(!is.na(DESIGNATED)) %>%
  mutate(SE_Flag = if_else(is.na(SE_Flag), 0, SE_Flag))

stopifnot(eeptools::isid(checking_rows, c("GEOID_b")))

checking_rows %>%
  mutate(GEOID_b = as.character(GEOID_b)) %>%
  group_by(GEOID_b) %>%
  mutate(n = n_distinct(GEOID)) %>%
  filter(n > 1) %>%
  nrow() %>%
  equals(0) %>%
  stopifnot()
```

```{r}
# write data
ar_oz_acs %>%
  as.data.frame() %>%
  mutate(distance_boundary = as.numeric(distance_boundary)) %>%
  write_csv("../00_data/ar_oz.csv")
```

```{r}
cp_oz_acs %>%
  as.data.frame() %>%
  mutate(distance_boundary = as.numeric(distance_boundary)) %>%
  write_csv("../00_data/cp_oz.csv")
```
