==> 02_programs/02b_create_analysis_file.Rmd <==
---
title: "02a_create_analysis_file"
author: "Adelaide Currin, Haowei Wang, Jennah Gosciak, Yiping Zuo"
date: "2023-02-06"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
library(sf)
library(lubridate)
library(zoo)
library(DescTools)
```

# Load Data

```{r, include=F}
# download data
ar_oz <- read_csv("../00_data/ar_oz.csv")
cp_oz <- read_csv("../00_data/cp_oz.csv")

centroid_distances <- st_read("../03_output/centroid_distances.geojson") %>%
  as.data.frame() %>%
  select(GEOID_b, centroid_distance_boundary, DESIGNATED) %>%
  # flagging out ineligible census blocks
  mutate(status = case_when(
    is.na(DESIGNATED) ~ "Ineligible",
    DESIGNATED == TRUE ~ "Selected",
    DESIGNATED == FALSE ~ "Unselected, Eligible"
  )) %>%
  mutate(GEOID_b = as.numeric(GEOID_b)) %>%
  select(-DESIGNATED)
centroid_distances

# checking for uniqueness
eeptools::isid(centroid_distances, "GEOID_b") %>%
  stopifnot()
```

```{r}
# load varlist
acs_varlist <- readxl::read_excel("../00_data/acs_varlist.xlsx") %>%
  filter(include == 1)
acs_varlist

acs_vars <- acs_varlist$variable
names(acs_vars) <- acs_varlist$description
constructs <- c(
  "population", "median_earnings", "median_household_income", "median_gross_rent",
  "pct_white", "pct_higher_ed", "pct_rent", "pct_native_hc_covered", "pct_poverty", "minutes_commute_avg",
  "dec_score", "SE_Flag", "vacancyrate", "pctunder18", "pctover64"
)

urbanvars <- c("dec_score", "SE_Flag", "vacancyrate", "pctunder18", "pctover64")
```

```{r}
# load acs data
acs_oz <- read_csv("../00_data/acs_oz.csv")

acs_oz_vars <- acs_oz %>%
  select(GEOID, constructs, urbanvars)

acs_oz %>%
  group_by(DESIGNATED, LIC) %>%
  summarize(n = n())
```

```{r}
# attempt at adjusted differences for arrests
ar_oz2 <- ar_oz %>%
  mutate(quarter = case_when(
    arrest_date >= mdy("06-14-2017") & arrest_date < mdy("09-14-2017") ~ -4,
    arrest_date >= mdy("09-14-2017") & arrest_date < mdy("12-14-2017") ~ -3,
    arrest_date >= mdy("12-14-2017") & arrest_date < mdy("03-14-2018") ~ -2,
    arrest_date >= mdy("03-14-2018") & arrest_date < mdy("06-14-2018") ~ -1,
    arrest_date >= mdy("06-14-2018") & arrest_date < mdy("09-14-2018") ~ 0,
    arrest_date >= mdy("09-14-2018") & arrest_date < mdy("12-14-2018") ~ 1,
    arrest_date >= mdy("12-14-2018") & arrest_date < mdy("03-14-2019") ~ 2,
    arrest_date >= mdy("03-14-2019") & arrest_date < mdy("06-14-2019") ~ 3,
    arrest_date >= mdy("06-14-2019") & arrest_date < mdy("09-14-2019") ~ 4,
    arrest_date >= mdy("09-14-2019") & arrest_date < mdy("12-14-2019") ~ 5,
    arrest_date >= mdy("12-14-2019") & arrest_date < mdy("03-14-2020") ~ 6,
    arrest_date >= mdy("03-14-2020") & arrest_date < mdy("06-14-2020") ~ 7,
    arrest_date >= mdy("06-14-2020") & arrest_date < mdy("09-14-2020") ~ 8,
    arrest_date >= mdy("09-14-2020") & arrest_date < mdy("12-14-2020") ~ 9,
    arrest_date >= mdy("12-14-2020") & arrest_date < mdy("03-14-2021") ~ 10
  )) %>%
  # filter(DESIGNATED == T | (DESIGNATED == F & LIC == T)) %>%
  mutate(
    treatment = if_else(DESIGNATED == T, 1, 0),
    distance_boundary = if_else(DESIGNATED == F, distance_boundary * -1, distance_boundary)
  ) %>%
  # note: calculating this at the Census block level
  mutate(post = if_else(arrest_date >= mdy("04-01-2018"), 1, 0)) %>%
  group_by(
    GEOID_b, GEOID, quarter, pct_white, pct_higher_ed, pct_rent, pct_native_hc_covered,
    pct_poverty, minutes_commute_avg, dec_score, SE_Flag, vacancyrate, pctunder18, pctover64
  ) %>%
  rename(
    violent_ellen_ar = violent_ellen,
    property_ellen_ar = property_ellen,
    distance_boundary_ar = distance_boundary
  ) %>%
  summarize(
    violent_ellen_ar = sum(violent_ellen_ar, na.rm = T),
    property_ellen_ar = sum(property_ellen_ar, na.rm = T),
    distance_boundary_ar = mean(distance_boundary_ar, na.rm = T)
  ) %>%
  mutate(SE_Flag = if_else(is.na(SE_Flag), 0, SE_Flag)) %>%
  filter(!is.na(quarter))

ar_oz2 <- ar_oz2
bind_rows(ar_oz2 %>%
  ungroup() %>%
  filter(quarter >= 0) %>%
  summarize(
    violent_ellen_ar = sum(violent_ellen_ar),
    property_ellen_ar = sum(property_ellen_ar)
  ) %>%
  mutate(quarter = 100))

stopifnot(eeptools::isid(ar_oz2, c("GEOID_b", "quarter")))
```

```{r}
# attempt at adjusted differences for complaints
cp_oz2 <- cp_oz %>%
  mutate(quarter = case_when(
    cmplnt_fr_dt >= mdy("06-14-2017") & cmplnt_fr_dt < mdy("09-14-2017") ~ -4,
    cmplnt_fr_dt >= mdy("09-14-2017") & cmplnt_fr_dt < mdy("12-14-2017") ~ -3,
    cmplnt_fr_dt >= mdy("12-14-2017") & cmplnt_fr_dt < mdy("03-14-2018") ~ -2,
    cmplnt_fr_dt >= mdy("03-14-2018") & cmplnt_fr_dt < mdy("06-14-2018") ~ -1,
    cmplnt_fr_dt >= mdy("06-14-2018") & cmplnt_fr_dt < mdy("09-14-2018") ~ 0,
    cmplnt_fr_dt >= mdy("09-14-2018") & cmplnt_fr_dt < mdy("12-14-2018") ~ 1,
    cmplnt_fr_dt >= mdy("12-14-2018") & cmplnt_fr_dt < mdy("03-14-2019") ~ 2,
    cmplnt_fr_dt >= mdy("03-14-2019") & cmplnt_fr_dt < mdy("06-14-2019") ~ 3,
    cmplnt_fr_dt >= mdy("06-14-2019") & cmplnt_fr_dt < mdy("09-14-2019") ~ 4,
    cmplnt_fr_dt >= mdy("09-14-2019") & cmplnt_fr_dt < mdy("12-14-2019") ~ 5,
    cmplnt_fr_dt >= mdy("12-14-2019") & cmplnt_fr_dt < mdy("03-14-2020") ~ 6,
    cmplnt_fr_dt >= mdy("03-14-2020") & cmplnt_fr_dt < mdy("06-14-2020") ~ 7,
    cmplnt_fr_dt >= mdy("06-14-2020") & cmplnt_fr_dt < mdy("09-14-2020") ~ 8,
    cmplnt_fr_dt >= mdy("09-14-2020") & cmplnt_fr_dt < mdy("12-14-2020") ~ 9,
    cmplnt_fr_dt >= mdy("12-14-2020") & cmplnt_fr_dt < mdy("03-14-2021") ~ 10
  )) %>%
  # filter(DESIGNATED == T | (DESIGNATED == F & LIC == T)) %>%
  mutate(
    treatment = if_else(DESIGNATED == T, 1, 0),
    distance_boundary = if_else(DESIGNATED == F, distance_boundary * -1, distance_boundary)
  ) %>%
  group_by(
    GEOID_b, GEOID, quarter, pct_white, pct_higher_ed, pct_rent, pct_native_hc_covered,
    pct_poverty, minutes_commute_avg, dec_score, SE_Flag, vacancyrate, pctunder18, pctover64
  ) %>%
  rename(
    violent_ellen_cp = violent_ellen,
    property_ellen_cp = property_ellen,
    distance_boundary_cp = distance_boundary
  ) %>%
  summarize(
    violent_ellen_cp = sum(violent_ellen_cp, na.rm = T),
    property_ellen_cp = sum(property_ellen_cp, na.rm = T),
    distance_boundary_cp = mean(distance_boundary_cp, na.rm = T)
  ) %>%
  mutate(SE_Flag = if_else(is.na(SE_Flag), 0, SE_Flag)) %>%
  filter(!is.na(quarter))

cp_oz2 <- cp_oz2
bind_rows(cp_oz2 %>%
  ungroup() %>%
  filter(quarter >= 0) %>%
  summarize(
    violent_ellen_cp = sum(violent_ellen_cp),
    property_ellen_cp = sum(property_ellen_cp)
  ) %>%
  mutate(quarter = 100))

stopifnot(eeptools::isid(cp_oz2, c("GEOID_b", "quarter")))
```

```{r}
# Winsorize
ar_oz2 <- ar_oz2 %>%
  mutate(
    violent_ellen_winsorize_ar = Winsorize(violent_ellen_ar, probs = c(0.01, 0.99)),
    property_ellen_winsorize_ar = Winsorize(property_ellen_ar, probs = c(0.01, 0.99))
  )

cp_oz2 <- cp_oz2 %>%
  mutate(
    violent_ellen_winsorize_cp = Winsorize(violent_ellen_cp, probs = c(0.01, 0.99)),
    property_ellen_winsorize_cp = Winsorize(property_ellen_cp, probs = c(0.01, 0.99))
  )
```


```{r}
# Create one file
ar_oz3 <- ar_oz2 %>%
  ungroup() %>%
  select(GEOID_b, GEOID, quarter, starts_with("violent_ellen"), starts_with("property_ellen"))

cp_oz3 <- cp_oz2 %>%
  ungroup() %>%
  select(GEOID_b, GEOID, quarter, starts_with("violent_ellen"), starts_with("property_ellen"))
cp_oz3
```
```{r}
# WITH remerging ACS but keeping original as well (so covaraite.x and covariate.y)
analysis_file <- full_join(ar_oz3, cp_oz3,
  by = c("GEOID_b", "quarter", "GEOID")
) %>%
  arrange(quarter) %>%
  mutate(quarter = str_replace(as.character(quarter), "\\-", "min")) %>%
  pivot_wider(
    id_cols = c(GEOID_b, GEOID),
    names_from = quarter,
    values_from = c(starts_with("violent_ellen"), starts_with("property_ellen"))
  ) %>%
  full_join(centroid_distances, by = "GEOID_b") %>%
  left_join(acs_oz_vars, by = "GEOID") %>%
  mutate(across(starts_with("violent_ellen"), ~ if_else(is.na(.), 0, .))) %>%
  mutate(across(starts_with("property_ellen"), ~ if_else(is.na(.), 0, .))) %>%
  # make distance negative if inside the boundary
  mutate(centroid_distance_boundary = if_else(status == "Selected", 
                                              centroid_distance_boundary, -1 * centroid_distance_boundary))

eeptools::isid(analysis_file, "GEOID_b") %>%
  stopifnot()
```

```{r}
# file for running with fixed effects
analysis_file_fe <- full_join(ar_oz3, cp_oz3,
  by = c("GEOID_b", "quarter", "GEOID")
) %>%
  arrange(quarter) %>%
  mutate(year = case_when(
    quarter < 0 ~ "year-1",
    quarter < 4 ~ "year0",
    quarter < 8 ~ "year1",
    TRUE ~ NA_character_
  )) %>%
  pivot_wider(
    id_cols = c(GEOID_b, GEOID, quarter),
    names_from = year,
    values_from = c(starts_with("violent_ellen"), starts_with("property_ellen"))
  ) %>%
  mutate(across(`violent_ellen_ar_year-1`:property_ellen_winsorize_cp_year1, ~ if_else(is.na(.), 0, .))) %>%
  full_join(centroid_distances, by = "GEOID_b") %>%
  left_join(acs_oz_vars, by = "GEOID") %>%
  mutate(across(starts_with("violent_ellen"), ~ if_else(is.na(.), 0, .))) %>%
  mutate(across(starts_with("property_ellen"), ~ if_else(is.na(.), 0, .))) %>%
  # make distance negative if outside the boundary
  mutate(centroid_distance_boundary = if_else(status == "Selected", centroid_distance_boundary,
    -1 * centroid_distance_boundary
  ))

analysis_file_fe

# analysis file with geom data

cb <- st_read("https://data.cityofnewyork.us/resource/twhy-dzjp.geojson?$limit=1000000") %>%
  st_transform(2263) %>%
  mutate(
    county_code = case_when(
      boro_code == 1 ~ "36061",
      boro_code == 2 ~ "36005",
      boro_code == 3 ~ "36047",
      boro_code == 4 ~ "36081",
      boro_code == 5 ~ "36085"
    ),
    GEOID = str_c(county_code, ct2010),
    GEOID_b = str_c(county_code, ct2010, cb2010)
  ) %>%
  mutate(geometry = st_make_valid(geometry)) %>%
  mutate(GEOID_b = as.numeric(GEOID_b)) %>%
  mutate(GEOID = as.numeric(GEOID))


analysis_file_geom <- left_join(analysis_file, cb, "GEOID_b" = "GEOID_b",
                                "GEOID" = "GEOID")
```


```{r}
# check missingness
unique_geoid_missing <- analysis_file %>%
  filter(is.na(pct_white)) %>%
  distinct(GEOID)

acs_oz_vars %>%
  inner_join(unique_geoid_missing, by = "GEOID") %>%
  filter(!is.na(pct_white)) %>%
  nrow() %>%
  magrittr::equals(0) %>%
  stopifnot()
```


```{r}
# write to csv
analysis_file

analysis_file %>%
  write_csv("../00_data/analysis_file.csv")

analysis_file_fe %>%
  write_csv("../00_data/analysis_file_fe.csv")

analysis_file_geom %>%
  st_write("../00_data/analysis_file_geom.geojson", delete_dsn = T)
```
