==> 02_programs/03_report_figures.Rmd <==
---
title: "Report Figures"
author: "Adelaide Currin, Haowei Wang, Jennah Gosciak, Yiping Zuo"
date: "2022-12-04"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
library(sf)
library(lubridate)
library(zoo)
library("lmtest")
library("sandwich")
library("wesanderson")
```

# Generates figures for each report

```{r, include=F}
# download data
ar_oz <- read_csv("../00_data/ar_oz.csv")

cp_oz <- read_csv("../00_data/cp_oz.csv")

ar_oz %>%
  head()

# load varlist
acs_varlist <- readxl::read_excel("../00_data/acs_varlist.xlsx") %>%
  filter(include == 1)
acs_varlist

acs_vars <- acs_varlist$variable
names(acs_vars) <- acs_varlist$description

constructs <- c(
  "pct_white", "pct_higher_ed",
  "pct_rent", "pct_native_hc_covered", "pct_poverty", "minutes_commute_avg"
)
urbanvars <- c("dec_score", "SE_Flag", "vacancyrate", "pctunder18", "pctover64")
```

```{r}
# create month, date aggregation
ar_oz <- ar_oz %>%
  mutate(arrest_month_yr = as.Date(
    str_c(month(arrest_date), 1, year(arrest_date), sep = "-"),
    "%m-%d-%Y"
  )) %>%
  mutate(year = year(arrest_month_yr))

cp_oz <- cp_oz %>%
  mutate(cp_month_yr = as.Date(
    str_c(month(cmplnt_fr_dt), 1, year(cmplnt_fr_dt), sep = "-"),
    "%m-%d-%Y"
  )) %>%
  mutate(year = year(cp_month_yr))

cp_oz <- cp_oz %>%
  mutate(cp_month_yr = as.Date(
    str_c(month(cmplnt_fr_dt), 1, year(cmplnt_fr_dt), sep = "-"),
    "%m-%d-%Y"
  )) %>%
  mutate(year = year(cp_month_yr))
```


```{r}
# average crime incidents by Census tract
ar_oz %>%
  filter(arrest_date < mdy("06-14-2018") & arrest_date >= mdy("06-14-2017")) %>%
  group_by(GEOID, DESIGNATED, arrest_month_yr) %>%
  summarize(n = n()) %>%
  group_by(DESIGNATED) %>%
  summarize(avg = mean(n))

cp_oz %>%
  filter(cmplnt_fr_dt < mdy("06-14-2018") & cmplnt_fr_dt >= mdy("06-14-2017")) %>%
  group_by(GEOID, DESIGNATED, cp_month_yr) %>%
  summarize(n = n()) %>%
  group_by(DESIGNATED) %>%
  summarize(avg = mean(n))
```

## Unadjusted and adjusted differences

```{r}
# attempt at adjusted differences for arrests
ar_oz_adj_diff <- ar_oz %>%
  # note: calculating this at the Census block level
  group_by(
    GEOID_b, arrest_month_yr, DESIGNATED, pct_white, pct_higher_ed, pct_rent, pct_native_hc_covered,
    pct_poverty, dec_score, SE_Flag, vacancyrate, pctunder18, pctover64,
  ) %>%
  summarize(across(c(violent_ellen, property_ellen), ~ sum(., na.rm = T))) %>%
  filter(!is.na(DESIGNATED)) %>%
  mutate(SE_Flag = if_else(is.na(SE_Flag), 0, SE_Flag))

ar_oz_adj_diff %>%
  select(DESIGNATED) %>%
  unique()
```

```{r}
# attempt at adjusted differences for complaints
cp_oz_adj_diff <- cp_oz %>%
  # note: calculating this at the Census block level
  group_by(
    GEOID_b, cp_month_yr, DESIGNATED, pct_white, pct_higher_ed, pct_rent, pct_native_hc_covered,
    pct_poverty, dec_score, SE_Flag, vacancyrate, pctunder18, pctover64
  ) %>%
  filter(!is.na(DESIGNATED)) %>%
  summarize(across(c(violent_ellen, property_ellen), ~ sum(., na.rm = T))) %>%
  mutate(SE_Flag = if_else(is.na(SE_Flag), 0, SE_Flag))

nrow(cp_oz_adj_diff)
```

```{r}
labs_ar <- c("Violent crime arrests", "Property crime arrests")
outcomes <- c("violent_ellen", "property_ellen")

labs_cp <- c("Violent crime complaints", "Property crime complaints")

create_plots <- function(df, outcomes, labs, fname, date_var) {
  # unadjusted plot
  df %>%
    rename_at(date_var, ~"date") %>%
    ungroup() %>%
    pivot_longer(c(violent_ellen, property_ellen)) %>%
    mutate(
      name = factor(labs[match(name, outcomes)],
        levels = labs,
        ordered = T
      ),
      desig_lab = factor(if_else(DESIGNATED == T, "Designated", "Eligible, not designated"),
        levels = c("Designated", "Eligible, not designated"),
        ordered = T
      )
    ) %>%
    ggplot() +
    stat_summary(aes(date, value, color = desig_lab),
      fun = "mean", geom = "line"
    ) +
    geom_vline(
      xintercept = mdy("04-01-2018"), linetype = "dashed",
      color = "gray"
    ) +
    theme_classic() +
    labs(
      y = "Number of incidents per block",
      x = "Month-year",
      color = "Designation"
    ) +
    scale_color_manual(values = wes_palette("Darjeeling1")) +
    facet_wrap(~name, scales = "free")

  ggsave(str_glue("../03_output/unadjusted_{fname}.png"), height = 6, width = 10)
}

create_plots(ar_oz_adj_diff, outcomes, labs_ar, "arrests", "arrest_month_yr")
create_plots(cp_oz_adj_diff, outcomes, labs_cp, "complaints", "cp_month_yr")
```
```{r}
outcomes <- c("resid_violent", "resid_property")

create_adjusted_plots <- function(df, outcomes, labs, fname, date_var) {
  # ideal model
  lmviolent <- lm(violent_ellen ~ pct_white + pct_higher_ed + pct_rent +
    pct_native_hc_covered + pct_poverty + dec_score + SE_Flag + vacancyrate + pctunder18 +
    pctover64, df)

  lmproperty <- lm(property_ellen ~ pct_white + pct_higher_ed + pct_rent +
    pct_native_hc_covered + pct_poverty + dec_score + SE_Flag + vacancyrate + pctunder18 + pctover64, df)
  modeldata <- cbind(df[rownames(lmviolent$model), ],
    "resid_violent" = lmviolent$residuals,
    "resid_property" = lmproperty$residuals
  ) %>%
    pivot_longer(c("resid_violent", "resid_property"))

  modeldata %>%
    nrow() %>%
    print()

  plot <- modeldata %>%
    ungroup() %>%
    rename_at(date_var, ~"date") %>%
    mutate(
      name = factor(labs[match(name, outcomes)],
        levels = labs,
        ordered = T
      ),
      desig_lab = factor(if_else(DESIGNATED == T, "Designated", "Eligible, not designated"),
        levels = c("Designated", "Eligible, not designated"),
        ordered = T
      )
    ) %>%
    ggplot() +
    stat_summary(aes(date, value, color = desig_lab),
      fun = "mean", geom = "line"
    ) +
    geom_vline(
      xintercept = mdy("04-01-2018"), linetype = "dashed",
      color = "gray"
    ) +
    ylim(-0.3, .3) +
    scale_color_manual(values = wes_palette("Darjeeling1")) +
    theme_classic() +
    labs(
      y = "Residual of incidents per block",
      x = "Month-year",
      color = "Designation"
    ) +
    facet_wrap(~name, scales = "free")

  ggsave(str_glue("../03_output/adjusted_{fname}.png"), height = 6, width = 10)
}

create_adjusted_plots(ar_oz_adj_diff, outcomes, labs_ar, "arrests", "arrest_month_yr")
create_adjusted_plots(cp_oz_adj_diff, outcomes, labs_cp, "complaints", "cp_month_yr")
```

## Descriptive stats and missing rates

```{r}
acs_oz <- read_csv("../00_data/acs_oz.csv")
acs_oz_bg <- read_csv("../00_data/acs_bg.csv")

acs_oz %>%
  group_by(DESIGNATED) %>%
  summarize(n = n()) %>%
  ungroup() %>%
  mutate(sum = cumsum(n))

acs_oz %>%
  filter(LIC == T) %>%
  group_by(DESIGNATED, LIC) %>%
  summarize(n = n()) %>%
  ungroup() %>%
  mutate(
    sum = cumsum(n),
    per_total = n / 1448,
    sum_per_total = sum / 1448
  )
```

```{r}
acs_oz %>%
  as.data.frame() %>%
  filter(!is.na(DESIGNATED)) %>%
  group_by(DESIGNATED) %>%
  # create missing flag
  # duplicate at the block group level
  # summary statistics of missingness
  # really only care about the ones that are close to the boundary
  summarize(across(population:pctover64, mean, na.rm = T)) %>%
  pivot_longer(cols = c(population:pctover64)) %>%
  mutate(
    value = round(value, 2),
    designated_desc = if_else(DESIGNATED == TRUE, "designated", "eligible")
  ) %>%
  pivot_wider(id_cols = "name", names_from = "designated_desc", values_from = "value") %>%
  select(name, designated, eligible) %>%
  write_csv("../03_output/desc_means.csv", na = "")

acs_oz %>%
  as.data.frame() %>%
  filter(!is.na(DESIGNATED)) %>%
  group_by(DESIGNATED) %>%
  # create missing flag
  # duplicate at the block group level
  # summary statistics of missingness
  # really only care about the ones that are close to the boundary
  summarize(across(population:pct_poverty, ~ sum(!is.na(.)))) %>%
  pivot_longer(cols = c(population:pct_poverty)) %>%
  mutate(
    value = round(value, 2),
    designated_desc = if_else(DESIGNATED == TRUE, "designated", "eligible")
  ) %>%
  pivot_wider(id_cols = "name", names_from = "designated_desc", values_from = "value") %>%
  select(name, designated, eligible) %>%
  summarize(
    designated = max(designated),
    eligible = max(eligible)
  )
```

```{r}
acs_oz %>%
  as.data.frame() %>%
  group_by(DESIGNATED) %>%
  # create missing flag
  # duplicate at the block group level
  # summary statistics of missingness
  # really only care about the ones that are close to the boundary
  mutate(across(population:pct_poverty, ~ if_else(is.na(.), 1, 0))) %>%
  summarize(across(population:pct_poverty, mean, na.rm = T)) %>%
  pivot_longer(cols = c(population:pct_poverty)) %>%
  mutate(
    value = round(value, 2),
    designated_desc = if_else(DESIGNATED == TRUE, "designated", "eligible")
  ) %>%
  pivot_wider(id_cols = "name", names_from = "designated_desc", values_from = "value") %>%
  select(name, designated, eligible) %>%
  write_csv("../03_output/missing_rates.csv", na = "")
```

```{r}
acs_oz_bg %>%
  as.data.frame() %>%
  group_by(DESIGNATED) %>%
  # create missing flag
  # duplicate at the block group level
  # summary statistics of missingness
  # really only care about the ones that are close to the boundary
  summarize(across(population:pct_poverty, mean, na.rm = T)) %>%
  pivot_longer(cols = c(population:pct_poverty)) %>%
  mutate(
    value = round(value, 2),
    designated_desc = if_else(DESIGNATED == TRUE, "designated", "eligible")
  ) %>%
  pivot_wider(id_cols = "name", names_from = "designated_desc", values_from = "value") %>%
  select(name, designated, eligible) %>%
  write_csv("../03_output/desc_means_bg.csv", na = "")

acs_oz_bg %>%
  as.data.frame() %>%
  group_by(DESIGNATED) %>%
  # create missing flag
  # duplicate at the block group level
  # summary statistics of missingness
  # really only care about the ones that are close to the boundary
  mutate(across(population:pct_poverty, ~ if_else(is.na(.), 1, 0))) %>%
  summarize(across(population:pct_poverty, mean, na.rm = T)) %>%
  pivot_longer(cols = c(population:pct_poverty)) %>%
  mutate(
    value = round(value, 2),
    designated_desc = if_else(DESIGNATED == TRUE, "designated", "eligible")
  ) %>%
  pivot_wider(id_cols = "name", names_from = "designated_desc", values_from = "value") %>%
  select(name, designated, eligible) %>%
  write_csv("../03_output/missing_rates_bg.csv", na = "")
```

## Balance tests overall

```{r}
# load adjacent census blocks and block groups
acs_oz_clean <- acs_oz %>%
  as.data.frame() %>%
  mutate(treatment = if_else(DESIGNATED == TRUE, 1, 0)) %>%
  filter(!is.na(DESIGNATED)) %>% 
  assertr::verify(!is.na(DESIGNATED)) %>%
  mutate(SE_Flag = if_else(is.na(SE_Flag), 0, SE_Flag))
```

```{r}
test_baseline_significance <- function(outcome, print = FALSE,
                                       data) {
  print(outcome)
  mod <- lm(paste0(outcome, " ~ treatment"), data = data)

  if (print == TRUE) {
    summary(mod) %>%
      print()
  }

  vals <- coeftest(mod, vcov = vcovHC(mod, type = "HC0"))
  print(vals)
  int <- vals[1, 1]
  treat_est <- vals[2, 1]
  treat_se <- vals[2, 2]
  treat_pval <- vals[2, 4]
  asterisk <- case_when(
    treat_pval < 0.01 ~ "***",
    treat_pval < 0.05 ~ "**",
    treat_pval < 0.1 ~ "*",
    TRUE ~ NA_character_
  )

  if (str_detect(outcome, "pct") | str_detect(outcome, "rate") |
    str_detect(outcome, "SE_Flag")) {
    treat_est <- treat_est * 100
    int <- int * 100
    treat_se <- treat_se * 100
  }
  treat_obs <- data %>%
    filter(treatment == 1) %>%
    nrow()
  comp_obs <- data %>%
    filter(treatment == 0) %>%
    nrow()
  data.frame(
    "var" = outcome,
    "mean_1" = round(int + treat_est, 2),
    "mean_0" = round(int, 2),
    "difference" = if_else(!is.na(asterisk), paste0(round(treat_est, 2), asterisk),
      as.character(round(treat_est, 2))
    ),
    "se" = round(treat_se, 2),
    "pval" = round(treat_pval, 4),
    "N" = nobs(mod),
    "treat_n" = treat_obs,
    "comp_n" = comp_obs
  )
}
```


```{r}
diff_results <- map_dfr(c(
  names(acs_vars),
  constructs, urbanvars
), ~ test_baseline_significance(., data = acs_oz_clean))

diff_results
diff_results %>%
  write_csv("../03_output/balance_tests.csv")
```
