==> 02_programs/04_analysis.Rmd <==
---
title: "Report Figures"
author: "Adelaide Currin, Haowei Wang, Jennah Gosciak, Yiping Zuo"
date: "2023-01-23"
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
library(sf)
library(lubridate)
library(zoo)
library(lmtest)
library(sandwich)
library(wesanderson)
library(rdrobust)
library(stargazer)
library(modelsummary)
library(kableExtra)
library(gt)

EXCLUDE_INELIGIBLE <- T
file_suffix <- "_keep_ineligible"
```

```{r}
# load centroid pairs
centroid_distance_pairs <- read_csv("../03_output/centroid_distance_pairs.csv")
centroid_distance_pairs
```


# Load Data

```{r, include=F}
# download data
analysis_file <- read_csv("../00_data/analysis_file.csv") %>%
  mutate(treatment = if_else(status == "Selected", 1, 0)) %>%
  mutate(treatment2 = case_when(
    status == "Selected" ~ 1,
    GEOID_b %in% centroid_distance_pairs$dist_pair ~ 0,
    TRUE ~ NA_real_
  ))

```

```{r}
analysis_file %>%
  mutate(
    missing_SE = is.na(SE_Flag),
    missing_pop = is.na(population)
  ) %>%
  filter(abs(centroid_distance_boundary) <= 200 & status == "Ineligible") %>%
  group_by(missing_pop) %>%
  summarize(n = n())

analysis_file %>%
  filter(is.na(population)) %>%
  select(vacancyrate, "pctunder18", population, GEOID)
```

```{r}
analysis_file_fe <- read_csv("../00_data/analysis_file_fe.csv") %>%
  mutate(
    quarter1 = if_else(quarter %in% c(-4, 0, 4), 1, 0),
    quarter2 = if_else(quarter %in% c(-3, 1, 5), 1, 0),
    quarter3 = if_else(quarter %in% c(-2, 2, 6), 1, 0),
    quarter4 = if_else(quarter %in% c(-1, 3, 7), 1, 0)
  )

analysis_file_fe %>% 
  group_by(GEOID_b) %>% 
  summarize(n = n())

if (EXCLUDE_INELIGIBLE == T) {
  analysis_file <- analysis_file %>%
    filter(status != "Ineligible")

  analysis_file_fe <- analysis_file_fe %>%
    filter(status != "Ineligible")

  file_suffix <- NA
}
```
```{r}
analysis_file %>%
  head()

analysis_file %>%
  group_by(treatment, treatment2) %>%
  summarize(n = n())
```

```{r}
# load varlist
acs_varlist <- readxl::read_excel("../00_data/acs_varlist.xlsx") %>%
  filter(include == 1)
acs_varlist

acs_vars <- acs_varlist$variable
names(acs_vars) <- acs_varlist$description

constructs <- c(
  "population", "median_earnings", "median_household_income", "median_gross_rent",
  "pct_white", "pct_higher_ed", "pct_rent", "pct_native_hc_covered", "pct_poverty",
  "dec_score", "SE_Flag", "vacancyrate", "pctunder18", "pctover64"
)
```

## Helper Functions

```{r}
tidy.rdrobust <- function(outcome, x, outcome_name, p = 1, c = 0, covs = NULL, ...) {
  if (!is.null(covs)) {
    object <- rdrobust(outcome, x,
      c = c, covs = covs,
      masspoints = "adjust", p = p,
      ...
    )

    covariate_lab <- "covariates"
  } else {
    object <- rdrobust(outcome, x,
      c = c,
      masspoints = "adjust", p = p,
      ...
    )

    covariate_lab <- "no covariates"
  }

  ret <- data.frame(
    outcome = outcome_name,
    term = row.names(object$coef),
    order = p,
    covs = covariate_lab,
    estimate = object$coef[, 1],
    std.error = object$se[, 1],
    statistic = object$z[, 1],
    p.value = object$pv[, 1],
    conf.low = object$ci[, 1],
    conf.high = object$ci[, 2],
    nobs.left = object$N[1],
    nobs.right = object$N[2],
    nobs.effective.left = object$N_h[1],
    nobs.effective.right = object$N_h[2],
    cutoff = object$c,
    order.regression = object$p,
    order.bias = object$q,
    kernel = object$kernel,
    bwselect = object$bwselect,
    bandwidth.h = str_c("[", object$bws[1, 1], ",", object$bws[1, 2], "]"),
    bandwidth.b = str_c("[", object$bws[2, 1], ",", object$bws[2, 2], "]")
  )
  row.names(ret) <- NULL
  ret
}
```

```{r}
## Baseline Checks
run_baseline_checks <- function(data, outcome, bandwidth=NA) {
  # drop missing covariate values
  if (!is.na(bandwidth)) {
    data <- data %>%
      filter(abs(centroid_distance_boundary) < bandwidth)
  }

  mod <- lm(paste0(outcome, "~ treatment"), data = data)

  vals <- coeftest(mod, vcov = vcovHC(mod, type = "HC0"), cluster = ~GEOID)
  int <- vals[1, 1]
  treat_est <- vals[2, 1]
  treat_se <- vals[2, 2]
  treat_pval <- vals[2, 4]
  asterisk <- case_when(
    treat_pval < 0.01 ~ "***",
    treat_pval < 0.05 ~ "**",
    treat_pval < 0.1 ~ "*",
    TRUE ~ NA_character_
  )

  if (str_detect(outcome, "pct") | str_detect(outcome, "rate") |
    str_detect(outcome, "SE_Flag")) {
    treat_est <- treat_est * 100
    int <- int * 100
    treat_se <- treat_se * 100
  }

  treat_obs <- data %>%
    filter(treatment == 1) %>%
    nrow()

  comp_obs <- data %>%
    filter(treatment == 0) %>%
    nrow()

  mean1 <- round(int + treat_est, 2)
  treat_est <- round(treat_est, 2)
  treat_se <- round(treat_se, 2)

  treat_est <- if_else(str_detect(outcome, "median") | outcome %in% c("population"),
    formatC(round(treat_est), mode = "integer", big.mark = ","), as.character(treat_est)
  )

  treat_se <- if_else(str_detect(outcome, "median") | outcome %in% c("population"),
    formatC(round(treat_se), big.mark = ",", mode = "integer"), as.character(round(treat_se, 2))
  )

  difference <- if_else(!is.na(asterisk), paste0(treat_est, asterisk), treat_est)

  data.frame(
    "var" = outcome,
    "mean_1" = mean1,
    "mean_0" = round(int, 2),
    "difference" = difference,
    "se" = treat_se,
    "pval" = round(treat_pval, 4),
    "N" = nobs(mod),
    "treat_n" = treat_obs,
    "comp_n" = comp_obs,
    "bandwidth" = bandwidth
  )
}
```


## BASELINE EQUIVALENCE

```{r, warnings=F}
baseline_data <- analysis_file %>%
  select(GEOID_b, constructs, centroid_distance_boundary, status, treatment, treatment2) %>%
  distinct()

stopifnot(eeptools::isid(analysis_file, "GEOID_b"))
stopifnot(eeptools::isid(baseline_data, "GEOID_b"))
```

```{r}
# balance test data
bequiv <- map_dfr(
  cross2(
    constructs,
    c(NA, seq(264, 1056, 264))
  ),
  ~ run_baseline_checks(baseline_data, outcome = .[[1]], bandwidth = .[[2]])
)

n_vals <- bequiv %>%
  select(var, N, bandwidth) %>%
  group_by(bandwidth) %>%
  summarize(N = formatC(max(N), big.mark = ",")) %>%
  mutate(var = "Total") %>%
  pivot_wider(id_cols = c("var"), names_from = "bandwidth", values_from = "N")
n_vals
```

```{r}
bequiv %>%
  select(var, N, bandwidth) %>%
  group_by(bandwidth) %>%
  mutate(N_new = formatC(max(N), big.mark = ",")) %>%
  arrange(bandwidth, desc(N))

bequiv %>%
  mutate(difference = str_c(difference, "\n(", se, ")")) %>%
  select(var, difference, bandwidth) %>%
  pivot_wider(id_cols = c("var"), names_from = "bandwidth", values_from = "difference") %>%
  bind_rows(n_vals) %>%
  write_csv(str_glue("../03_output/baseline_form_lm {Sys.Date()}{file_suffix}.csv"))
```

```{r}
# balance test data paired
bequiv <- map_dfr(
  cross2(
    constructs,
    seq(264, 1056, 264)
  ),
  ~ run_baseline_checks(baseline_data %>%
    select(-treatment, treatment = treatment2), outcome = .[[1]], bandwidth = .[[2]])
)

n_vals <- bequiv %>%
  select(var, N, bandwidth) %>%
  group_by(bandwidth) %>%
  summarize(N = formatC(max(N), big.mark = ",")) %>%
  mutate(var = "Total") %>%
  pivot_wider(id_cols = c("var"), names_from = "bandwidth", values_from = "N")

bequiv %>%
  mutate(difference = str_c(difference, "\n(", se, ")")) %>%
  select(var, difference, bandwidth) %>%
  pivot_wider(id_cols = c("var"), names_from = "bandwidth", values_from = "difference") %>%
  bind_rows(n_vals) %>%
  write_csv(str_glue("../03_output/baseline_form_lm_paired {Sys.Date()}{file_suffix}.csv"))
```

## RESULTS

```{r}
run_model_set <- function(outcome_data, running_var, outcomes, p, covs, c=0) {
  # vector to drop missing covariates
  keep_cov <- covs %>%
    is.na() %>%
    rowSums() %>%
    magrittr::equals(0)

  # drop missing covariate values
  outcome_data_cov <- outcome_data[keep_cov, ]
  running_var_cov <- running_var[keep_cov]
  covs <- covs[keep_cov, ]

  return(map2_dfr(
    outcome_data, outcomes,
    ~ tidy.rdrobust(.x, running_var, .y, p = p, c = c)
  ) %>%
    bind_rows(map2_dfr(
      outcome_data_cov, outcomes,
      ~ tidy.rdrobust(.x, running_var_cov, .y, p = p, c = c, covs = covs)
    )))
}

run_model_set_cluster <- function(outcome_data, running_var, outcomes, p, covs, c=0, cluster) {
  # vector to drop missing covariates
  keep_cov <- covs %>%
    is.na() %>%
    rowSums() %>%
    magrittr::equals(0)

  # drop missing covariate values
  outcome_data_cov <- outcome_data[keep_cov, ]
  running_var_cov <- running_var[keep_cov]
  covs <- covs[keep_cov, ]
  censustracts_cov <- cluster[keep_cov]

  return(map2_dfr(
    outcome_data, outcomes,
    ~ tidy.rdrobust(.x, running_var, .y, p = p, c = c, cluster = cluster)
  ) %>%
    bind_rows(map2_dfr(
      outcome_data_cov, outcomes,
      ~ tidy.rdrobust(.x, running_var_cov, .y, p = p, covs = covs, c = c, cluster = censustracts_cov)
    )))
}

theory_run_model_set_cluster <- function(outcome_data, running_var, outcomes, p, covs, bandwidth, cluster) {
  # vector to drop missing covariates
  keep_cov <- covs %>%
    is.na() %>%
    rowSums() %>%
    magrittr::equals(0)

  # drop missing covariate values
  outcome_data_cov <- outcome_data[keep_cov, ]
  running_var_cov <- running_var[keep_cov]
  covs <- covs[keep_cov, ]
  censustracts_cov <- cluster[keep_cov]

  return(map2_dfr(
    outcome_data, outcomes,
    ~ tidy.rdrobust(.x, running_var, .y, p = p, h = bandwidth, b = bandwidth, cluster = cluster)
  ) %>%
    bind_rows(map2_dfr(
      outcome_data_cov, outcomes,
      ~ tidy.rdrobust(.x, running_var_cov, .y,
        p = p, covs = covs, h = bandwidth, b = bandwidth, cluster =
          censustracts_cov
      )
    )))
}

theory_run_model_set <- function(outcome_data, running_var, outcomes, p, covs, bandwidth) {
  # vector to drop missing covariates
  keep_cov <- covs %>%
    is.na() %>%
    rowSums() %>%
    magrittr::equals(0)

  # drop missing covariate values
  outcome_data_cov <- outcome_data[keep_cov, ]
  running_var_cov <- running_var[keep_cov]
  covs <- covs[keep_cov, ]

  return(map2_dfr(
    outcome_data, outcomes,
    ~ tidy.rdrobust(.x, running_var, .y, p = p, h = bandwidth, b = bandwidth)
  ) %>%
    bind_rows(map2_dfr(
      outcome_data_cov, outcomes,
      ~ tidy.rdrobust(.x, running_var_cov, .y,
        p = p, covs = covs, h = bandwidth, b = bandwidth)
    )))
}
```

## RESULTS SETUP

```{r}
# define data and outcomes
outcomes_ar <- map(
  c("violent_ellen_winsorize_ar_", "property_ellen_winsorize_ar_"),
  ~ str_c(., c("min4", "min3", "min2", "min1", 0, c(1:7, 10)))
) %>%
  unlist()

outcome_data_ar <- analysis_file[, outcomes_ar]
running_var <- analysis_file[["centroid_distance_boundary"]]
covariates <- analysis_file[, constructs]
censustracts <- analysis_file[["GEOID"]]

outcomes_cp <- map(
  c("violent_ellen_winsorize_cp_", "property_ellen_winsorize_cp_"),
  ~ str_c(., c("min4", "min3", "min2", "min1", 0, c(1:7, 10)))
) %>%
  unlist()
outcome_data_cp <- analysis_file[, outcomes_cp]
```

```{r}
output <- map_dfr(
  c(1, 4),
  ~ run_model_set_cluster(outcome_data_ar, running_var, outcomes_ar, p = ., 
                          covs =covariates, cluster= censustracts)
) %>%
  bind_rows(map_dfr(
    c(1, 4),
    ~ run_model_set_cluster(outcome_data_cp, running_var, outcomes_cp, p = ., 
                            covs =covariates, cluster= censustracts)
  ))

output <- output %>%
  separate(outcome, c("crime_type", "measurement", "winsorize", "arrest_or_complaint", "quarter")) %>%
  select(-c(measurement)) %>%
  select(crime_type, arrest_or_complaint, winsorize, term, order, covs, quarter, everything()) %>%
  arrange(crime_type, arrest_or_complaint, term, order, covs, quarter)

output %>%
  write_csv(str_glue("../03_output/results_{Sys.Date()}{file_suffix}.csv"))

output
output %>%
  filter(winsorize == "winsorize") %>%
  mutate(nobs = rowSums(select(., c(nobs.effective.left, nobs.effective.right)), na.rm = T)) %>%
  group_by(covs, crime_type, order) %>% 
  summarize(nobs = min(nobs))

output %>%
  filter(winsorize == "winsorize") %>%
  mutate(
    estimate_form = case_when(
      p.value < 0.01 ~ paste0(round(estimate, 2), "***"),
      p.value < 0.05 ~ paste0(round(estimate, 2), "**"),
      p.value < 0.1 ~ paste0(round(estimate, 2), "*"),
      TRUE ~ paste0(round(estimate, 2))
    ),
    coef_form = str_c(estimate_form, "\n(", round(std.error, 2), ")")
  ) %>%
  pivot_wider(
    id_cols = c("quarter", "order", "covs"),
    names_from = c("crime_type", "term", "arrest_or_complaint"),
    values_from = "coef_form"
  ) %>%
  select(c("quarter", "order", "covs"), matches("Conventional"), 
         matches("Bias"), matches("Robust")) %>%
  select(c("quarter", "order", "covs"), 
         matches("property"), matches("violent")) %>%
  write_csv(str_glue("../03_output/results_form_{Sys.Date()}{file_suffix}.csv"))
```

## THEORY-DRIVEN BANDWIDTHS

```{r}
bandwidththeory_lin <- map_dfr(
  seq(264, 792, 264),
  ~ theory_run_model_set_cluster(outcome_data_ar, running_var, outcomes_ar, p = 1, covariates, bandwidth = ., censustracts)
) %>%
  bind_rows(map_dfr(
    seq(264, 792, 264),
    ~ theory_run_model_set_cluster(outcome_data_cp, running_var, outcomes_cp, p = 1, covariates, bandwidth = ., censustracts)
  ))
```

```{r}
bandwidththeory_poly <- map_dfr(
  seq(264, 792, 264),
  ~ theory_run_model_set_cluster(outcome_data_ar, running_var, outcomes_ar,
    p = 4, covariates,
    bandwidth = ., censustracts
  )
) %>%
  bind_rows(map_dfr(
    seq(264, 792, 264),
    ~ theory_run_model_set_cluster(outcome_data_cp, running_var, outcomes_cp, p = 4, covariates, bandwidth = ., censustracts)
  ))
```


```{r}
# Formatting
bandwidththeory_all <- bind_rows(bandwidththeory_lin, bandwidththeory_poly)

bandwidththeory_all_formatted <- bandwidththeory_all %>%
  separate(outcome, c("crime_type", "measurement", "winsorize", "arrest_or_complaint", "quarter")) %>%
  select(-c(measurement)) %>%
  select(crime_type, arrest_or_complaint, winsorize, term, order, covs, quarter, everything()) %>%
  arrange(crime_type, arrest_or_complaint, term, order, covs, quarter)

bandwidththeory_all_formatted %>%
  write_csv(str_glue("../03_output/results_theory_{Sys.Date()}{file_suffix}.csv"))

bandwidththeory_all_formatted %>%
  filter(winsorize == "winsorize") %>%
  mutate(nobs = rowSums(select(., c(nobs.effective.left, nobs.effective.right)), na.rm = T)) %>%
  group_by(covs, order, crime_type) %>% 
  summarize(nobs = min(nobs))

bandwidththeory_all_formatted %>%
  filter(winsorize == "winsorize") %>%
  filter(bandwidth.h == "[528,528]") %>%
  mutate(
    estimate_form = case_when(
      p.value < 0.01 ~ paste0(round(estimate, 2), "***"),
      p.value < 0.05 ~ paste0(round(estimate, 2), "**"),
      p.value < 0.1 ~ paste0(round(estimate, 2), "*"),
      TRUE ~ paste0(round(estimate, 2))
    ),
    coef_form = str_c(estimate_form, "\n(", round(std.error, 2), ")")
  ) %>%
  pivot_wider(
    id_cols = c("quarter", "order", "covs"), names_from = c("crime_type", "term", "arrest_or_complaint"),
    values_from = "coef_form"
  ) %>%
  select(c("quarter", "order", "covs"), matches("_ar"), matches("_cp")) %>%
  select(c("quarter", "order", "covs"), matches("Conventional"), matches("Bias"), matches("Robust")) %>%
  select(c("quarter", "order", "covs"), matches("property"), matches("violent")) %>%
  write_csv(str_glue("../03_output/results_form_theory_{Sys.Date()}{file_suffix}.csv"))
```


## PLACEBO TESTS GEOGRAPHIC

```{r}
output <- map_dfr(
  c(-1056, -792, -528, -264),
  ~ run_model_set_cluster(outcome_data_ar, running_var, outcomes_ar, p = 4, c = ., covariates, censustracts)
) %>%
  bind_rows(map_dfr(
     c(-1056),
    ~ run_model_set_cluster(outcome_data_cp, running_var, outcomes_cp, p = 4, c = ., covariates, censustracts)
  ))

output <- output %>%
  separate(outcome, c("crime_type", "measurement", "winsorize", "arrest_or_complaint", "quarter")) %>%
  select(-c(measurement)) %>%
  select(crime_type, arrest_or_complaint, winsorize, term, order, covs, quarter, cutoff, everything()) %>%
  arrange(crime_type, arrest_or_complaint, term, order, covs, quarter)

output %>%
  write_csv(str_glue("../03_output/placebo_{Sys.Date()}{file_suffix}.csv"))

output
output %>%
  filter(winsorize == "winsorize") %>%
  mutate(nobs = rowSums(select(., c(nobs.left, nobs.right)), na.rm = T)) %>%
  distinct(nobs, covs, crime_type)

output %>%
  filter(winsorize == "winsorize") %>%
  mutate(
    estimate_form = case_when(
      p.value < 0.01 ~ paste0(round(estimate, 2), "***"),
      p.value < 0.05 ~ paste0(round(estimate, 2), "**"),
      p.value < 0.1 ~ paste0(round(estimate, 2), "*"),
      TRUE ~ paste0(round(estimate, 2))
    ),
    coef_form = str_c(estimate_form, "\n(", round(std.error, 2), ")")
  ) %>%
  pivot_wider(
    id_cols = c("quarter", "order", "covs"), names_from = c("crime_type", "term", "arrest_or_complaint", "cutoff"),
    values_from = "coef_form"
  ) %>%
  select(c("quarter", "order", "covs"), matches("Robust")) %>%
  select(c("quarter", "order", "covs"), ends_with("-1056")) %>%
  select(c("quarter", "order", "covs"), matches("property"), matches("violent")) %>%
  write_csv(str_glue("../03_output/placebo_form_{Sys.Date()}{file_suffix}.csv")) 

output %>%
  filter(winsorize == "winsorize") %>%
  mutate(
    estimate_form = case_when(
      p.value < 0.01 ~ paste0(round(estimate, 2), "***"),
      p.value < 0.05 ~ paste0(round(estimate, 2), "**"),
      p.value < 0.1 ~ paste0(round(estimate, 2), "*"),
      TRUE ~ paste0(round(estimate, 2))
    ),
    coef_form = str_c(estimate_form, "\n(", round(std.error, 2), ")")
  ) %>%
  pivot_wider(
    id_cols = c("quarter", "order", "covs"), names_from = c("crime_type", "term", "arrest_or_complaint", "cutoff"),
    values_from = "coef_form"
  ) %>%
  select(c("quarter", "order", "covs"), 
         matches("Robust")) %>%
  select(c("quarter", "order", "covs"), 
         ends_with("-264"), 
         ends_with("-528"), ends_with("-792"), ends_with("-1056")) %>%
  select(c("quarter", "order", "covs"), matches("property"), matches("violent")) %>%
  write_csv(str_glue("../03_output/placebo_form_{Sys.Date()}{file_suffix}.csv"))
```

## FIXED EFFECTS
```{r}
# define data and outcomes
outcomes_ar <- map(
  c("violent_ellen_winsorize_ar_", 
    "property_ellen_winsorize_ar_"),
  ~ str_c(., c("year-1", "year0", "year1"))
) %>%
  unlist()

outcomes_ar <- outcomes_ar
outcomes_cp <- outcomes_cp
running_var <- analysis_file_fe[["centroid_distance_boundary"]]
outcome_data_ar <- analysis_file_fe[, outcomes_ar]
covariates <- analysis_file_fe[, c(constructs,
                                   str_c("quarter", 1:3))]
censustracts <- analysis_file_fe[["GEOID"]]

outcomes_cp <- map(c("violent_ellen_winsorize_cp_", 
                     "property_ellen_winsorize_cp_"), 
                   ~ str_c(., c("year-1", "year0", "year1"))) %>%
  unlist()
outcome_data_cp <- analysis_file_fe[, outcomes_cp]
```

```{r}
output_fe <- map_dfr(
  c(1, 4),
  ~ run_model_set(outcome_data_ar, running_var, 
                             outcomes_ar, p = ., covs=covariates)
  ) %>%
  bind_rows(map_dfr(
    c(1, 4),
    ~ run_model_set(outcome_data_cp, 
                    running_var,
                    outcomes_cp, p = ., covs=covariates)
  ))

output_fe <- output_fe %>%
  separate(outcome, c("crime_type", 
                      "measurement", 
                      "winsorize", 
                      "arrest_or_complaint", "year")) %>%
  select(-c(measurement)) %>%
  select(crime_type, arrest_or_complaint, 
         winsorize, term, 
         order, covs, year, everything()) %>%
  arrange(crime_type, arrest_or_complaint, 
          term, order, covs, year)

output_fe
output_fe %>%
  write_csv(str_glue("../03_output/results_fe_{Sys.Date()}{file_suffix}.csv"))
```

```{r}
output_fe %>%
  filter(winsorize == "winsorize") %>%
  mutate(nobs = rowSums(select(., c(nobs.effective.left, nobs.effective.right)), na.rm = T)) %>%
  group_by(covs, crime_type, order) %>% 
  summarize(nobs = min(nobs))

output_fe %>%
  filter(winsorize == "winsorize") %>%
  mutate(
    estimate_form = case_when(
      p.value < 0.01 ~ paste0(round(estimate, 2), "***"),
      p.value < 0.05 ~ paste0(round(estimate, 2), "**"),
      p.value < 0.1 ~ paste0(round(estimate, 2), "*"),
      TRUE ~ paste0(round(estimate, 2))
    ),
    coef_form = str_c(estimate_form, "\n(", round(std.error, 2), ")")
  ) %>%
  pivot_wider(
    id_cols = c("year", "order", "covs"), names_from = c("crime_type", "term", "arrest_or_complaint"),
    values_from = "coef_form"
  ) %>%
  select(c("year", "order", "covs"), 
         matches("Conventional"),
         matches("Bias"),
         matches("Robust")) %>%
  select(c("year", "order", "covs"), matches("property"), matches("violent")) %>%
  write_csv(str_glue("../03_output/results_form__fe_{Sys.Date()}{file_suffix}.csv"))
```

## PLOTS

```{r}
generate_rd_plot <- function(df, outcome_var, outcome_lab, p, covariates, type, time, h, ...) {
  df_filt <- df %>%
    rename(outcome = outcome_var) %>%
    filter(abs(centroid_distance_boundary) <= h)

  outcome_col <- df_filt %>%
    magrittr::extract2("outcome")

  running_var <- df_filt %>%
    magrittr::extract2("centroid_distance_boundary")

  covs <- df_filt[, covariates]

  # creating plots
  rdplot(outcome_col, running_var,
    p = p, y.label = str_glue("Average number of {type} per {time} (p={p})"),
    x.label = "distance", title = outcome_lab,
    covs = covs, y.lim = c(0, 1), h = h, ...
  )

  ggsave(str_glue("../03_output/rdplot_{outcome_var}_{p}.png"))
}
```


```{r}
walk2(c(1, 4), c(400), ~ generate_rd_plot(analysis_file, "violent_ellen_winsorize_cp_0", "Violent crime complaints, quarter 0", .x,
  constructs, "complaints", "quarter",
  h = .y
))

walk2(c(1, 4), c(400), ~ generate_rd_plot(analysis_file, "property_ellen_winsorize_cp_0", "Property crime complaints, quarter 0", .x,
  constructs, "complaints", "quarter",
  h = .y
))

walk2(c(1, 4), c(400), ~ generate_rd_plot(analysis_file, "violent_ellen_winsorize_cp_4", "Violent crime complaints, quarter 4", .x,
  constructs, "complaints", "quarter",
  h = .y
))

walk2(c(1, 4), c(400), ~ generate_rd_plot(analysis_file, "property_ellen_winsorize_cp_4", "Property crime complaints, quarter 4", .x,
  constructs, "complaints", "quarter",
  h = .y
))
```


```{r}
plot_data <- analysis_file %>%
  select(c(GEOID_b, status, centroid_distance_boundary, matches("violent_ellen_winsorize"))) %>%
  pivot_longer(matches("violent_ellen_winsorize")) %>%
  separate(name, c("crime_type", "method", "winsorize", "ar_cp", "quarter")) %>%
  filter(ar_cp == "ar") %>%
  mutate(treatment = if_else(status == "Selected", 1, 0))

plot_data %>%
  filter(abs(centroid_distance_boundary) < 193) %>%
  ggplot(aes(x = centroid_distance_boundary, y = value, group = status, color = status)) +
  geom_point(alpha = 0.25) +
  # geom_smooth(method="lm") +
  theme_classic() +
  facet_wrap(~quarter)
```
```{r}
points <- analysis_file %>%
  filter(abs(centroid_distance_boundary) < 193) %>%
  mutate(centroid_distance_boundary = round(centroid_distance_boundary, 1)) %>%
  group_by(centroid_distance_boundary) %>%
  summarize(across(matches("violent_ellen_winsorize"), ~ mean(., na.rm = T))) %>%
  pivot_longer(-centroid_distance_boundary)

points
```

```{r}
plot_data %>%
  filter(abs(centroid_distance_boundary) < 193) %>%
  ggplot(aes(x = centroid_distance_boundary, y = value, color = status)) +
  geom_point(data = points, aes(x = centroid_distance_boundary, y = value), alpha = .2, inherit.aes = F) +
  geom_smooth(method = "lm") +
  theme_classic() +
  facet_wrap(~quarter)
```
